add sortedby test, check length before loop

+++ models/articles.models.js
@@ -21,7 +21,7 @@ exports.fetchArticles = async () => {
         articles.created_at,
         articles.votes,
         articles.article_img_url,
-        COUNT(comments) AS comment_count
+        COUNT(comments)::INT AS comment_count
     FROM articles
     LEFT JOIN comments ON articles.article_id = comments.article_id
     GROUP BY articles.article_id
